+++ android/plugins/nopegl-nodes/src/main/kotlin/org/nopeforge/nopegl/specs/nodeType.kt
@@ -82,6 +82,8 @@ fun nodeType(name: String): NGLNodeType? {
         "Compute" -> NGLNodeType.COMPUTE
         "ComputeProgram" -> NGLNodeType.COMPUTEPROGRAM
         "CustomTexture" -> NGLNodeType.CUSTOMTEXTURE
+        "ColorFill" -> NGLNodeType.COLORFILL
+        "CustomFill" -> NGLNodeType.CUSTOMFILL
         "Draw" -> NGLNodeType.DRAW
         "DrawColor" -> NGLNodeType.DRAWCOLOR
         "DrawDisplace" -> NGLNodeType.DRAWDISPLACE
@@ -93,6 +95,7 @@ fun nodeType(name: String): NGLNodeType? {
         "DrawPath" -> NGLNodeType.DRAWPATH
         "DrawTexture" -> NGLNodeType.DRAWTEXTURE
         "DrawWaveform" -> NGLNodeType.DRAWWAVEFORM
+        "DrawRect" -> NGLNodeType.DRAWRECT
         "FilterAlpha" -> NGLNodeType.FILTERALPHA
         "FilterColorMap" -> NGLNodeType.FILTERCOLORMAP
         "FilterContrast" -> NGLNodeType.FILTERCONTRAST
@@ -107,6 +110,8 @@ fun nodeType(name: String): NGLNodeType? {
         "FastGaussianBlur" -> NGLNodeType.FASTGAUSSIANBLUR
         "FontFace" -> NGLNodeType.FONTFACE
         "GaussianBlur" -> NGLNodeType.GAUSSIANBLUR
+        "Gradient4Fill" -> NGLNodeType.GRADIENT4FILL
+        "GradientFill" -> NGLNodeType.GRADIENTFILL
         "Geometry" -> NGLNodeType.GEOMETRY
         "GraphicConfig" -> NGLNodeType.GRAPHICCONFIG
         "GridLayout" -> NGLNodeType.GRIDLAYOUT
@@ -133,6 +138,7 @@ fun nodeType(name: String): NGLNodeType? {
         "EvalVec3" -> NGLNodeType.EVALVEC3
         "EvalVec4" -> NGLNodeType.EVALVEC4
         "Media" -> NGLNodeType.MEDIA
+        "NoiseFill" -> NGLNodeType.NOISEFILL
         "NoiseFloat" -> NGLNodeType.NOISEFLOAT
         "NoiseVec2" -> NGLNodeType.NOISEVEC2
         "NoiseVec3" -> NGLNodeType.NOISEVEC3
@@ -152,6 +158,10 @@ fun nodeType(name: String): NGLNodeType? {
         "Scale" -> NGLNodeType.SCALE
         "Skew" -> NGLNodeType.SKEW
         "SmoothPath" -> NGLNodeType.SMOOTHPATH
+        "Stroke" -> NGLNodeType.STROKE
+        "StrokeGradient" -> NGLNodeType.STROKEGRADIENT
+        "StrokeGradient4" -> NGLNodeType.STROKEGRADIENT4
+        "TextureFill" -> NGLNodeType.TEXTUREFILL
         "Text" -> NGLNodeType.TEXT
         "TextEffect" -> NGLNodeType.TEXTEFFECT
         "Texture2D" -> NGLNodeType.TEXTURE2D